requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
transformers>=4.35.0
torch>=2.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.17.0
reportlab>=4.0.0

